Lexeme Table
Added a function to vector.c that allows for appending strings to the
vector.
Added two new vectors that store the character data for lexemetable.txt
and lexemelist.txt
lexemetable.txt's vector (aryLexTable) is generated fully and completely
now as far as I'm aware. (No doubt there are bugs left to sniff out).